Improvements
- Added ability to enable or disable individual modules in the web interface
 - The default setting for all modules is disabled
- Switched to using hls.js from npm package manager in build process
- Improved 2web.sh parallel processing
- Fixed bug in music2web causing every other track scanned to fail being
  scanned
- Simplified text in left and right page turn buttons for comic2web
- Refactored modules.php settings page to make the code more readable
- Improvements to music2web user output formatting
- Added supporting projects list to all man pages
- All 2web module nuke commands now remove related cache files
- nfo2web
 - nfo2web nuke now will now only remove nfo2web module content
 - moved update code into a function to cleanup interface code

+++ templates/stats.php
@@ -125,9 +125,24 @@ echo "		<span>";
 echo "			Free:$freeSpace";
 echo "		</span>";
 echo "	</div>";
+# check the status of the fortunes for drawing large or small widgets
+$fortuneEnabled = False;
+if ( file_exists("/etc/2web/fortuneStatus.cfg")){
+	$fortuneEnabled = True;
+}
+$weatherEnabled = False;
+if ( file_exists("/etc/2web/weather/homepageLocation.cfg")){
+	if ( file_exists($_SERVER['DOCUMENT_ROOT']."/weather.index")){
+		$weatherEnabled = True;
+	}
+}
 if ( file_exists("/etc/2web/fortuneStatus.cfg")){
 	echo "<a class='homeWeather' href='/fortune.php'>";
-	echo "<div class='inputCard'>";
+	if ($weatherEnabled){
+		echo "<div class='inputCard'>";
+	}else{
+		echo "<div class='listCard'>";
+	}
 	echo "<h3>Fortune</h3>";
 	echo "<div class='fortuneText'>";
 	echo "$todaysFortune";
@@ -136,11 +151,17 @@ if ( file_exists("/etc/2web/fortuneStatus.cfg")){
 	echo "</a>";
 }
 if ( file_exists("/etc/2web/weather/homepageLocation.cfg")){
-	echo "<a class='homeFortune' href='/weather/#".str_replace("\n","",file_get_contents("/etc/2web/weather/homepageLocation.cfg"))."'>";
-	echo "<div class='inputCard'>";
-	echo "$todaysWeather";
-	echo "</div>";
-	echo "</a>";
+	if ( file_exists($_SERVER['DOCUMENT_ROOT']."/weather.index")){
+		echo "<a class='homeFortune' href='/weather/#".str_replace("\n","",file_get_contents("/etc/2web/weather/homepageLocation.cfg"))."'>";
+		if ($fortuneEnabled){
+			echo "<div class='inputCard'>";
+		}else{
+			echo "<div class='listCard'>";
+		}
+		echo "$todaysWeather";
+		echo "</div>";
+		echo "</a>";
+	}
 }
 echo "</div>"
 ?>
